bird-report/requirements.txt
pandas==1.4.4
PyExecJS==1.5.1
Requests==2.30.0
